bug 106 (OS X: "Symbol not found: _dbadata" error when importing pymssql): Fix OS X build by modifying setup.py so that on OS X we don't attempt to link with the bundled FreeTDS *Linux* library.
The included FreeTDS Linux library in freetds/nix_64/lib doesn't work with OS
X, The setup.py erroneously tries to link with this library even on an OS X
system.

This commit changes the location searched for libraries on OS X. It will not
use a bundled library (we don't bundle one yet) and will instead link with a
library on the system from Homebrew, MacPorts, Fink, etc.

Fixes: http://code.google.com/p/pymssql/issues/detail?id=106