Modify get_row() to use the CPython tuple API
This drops the previous method of building up a row tuple and switches
to using the CPython API, which allows you to create a correctly sized
tuple at the beginning and simply fill it in. This appears to offer
around a 10% boost when fetching rows from a table where the data is
already in memory.

The CPU usage is still up around 100% so more optimisations should be
able to be made somewhere if some proper profiling was done.