Update Dockerfile
The previous dockerfile does not build correctly as it has not been updated in a while. It also uses a non-standard python image with a 14.04 Ubuntu base and adds in more pip packages than required to install pymssql. The changes here use the base 18.04 image, installs python3 from repositories, and installs cython via pip.