Use dbanydatecrack for TIME and DATETIME2.
TIME and DATETIME2 values can have precision varying between one
second (TIME(0), DATETIME2(0)) up to 100 nanoseconds (TIME(7),
DATETIME2(7)).  (In contrast, the older DATETIME type has precision of
1/300 second, or about 3.333 milliseconds.)

When a TIME or DATETIME2 value is retrieved from the database
(assuming we are using TDS 7.3 or later), it is converted into a
datetime.time or datetime.datetime object, respectively.  When doing
so, the precision of the original value should be retained if
possible.

To accomplish this, use the new dbanydatecrack function, instead of
first converting the value into a DATETIME and then using dbdatecrack.

(Unfortunately, the Python datetime.time and datetime.datetime types
provide one-microsecond precision at best.  So TIME(7) and
DATETIME2(7) values are still rounded to the nearest microsecond; but
this is better than first rounding to the nearest 1/300 second and
then rounding to the nearest millisecond.)

As a result of this change:

  >>> cursor.execute("SELECT CAST('12:34:56.789123' AS TIME)")
  >>> cursor.fetchall()
  [(datetime.time(12, 34, 56, 789123),)]
  >>> cursor.execute("SELECT CAST('2000-01-01 12:34:56.789123' AS DATETIME2)")
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 789123),)]

whereas previously:

  >>> cursor.execute("SELECT CAST('12:34:56.789123' AS TIME)")
  >>> cursor.fetchall()
  [(datetime.time(12, 34, 56, 790000),)]
  >>> cursor.execute("SELECT CAST('2000-01-01 12:34:56.789123' AS DATETIME2)")
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 790000),)]