Add MSSQLTimezone class.
This class can be used to represent a time zone with an arbitrary
fixed offset from UTC.

This class is essentially a simpler version of the datetime.timezone
class (which is part of the standard library since Python 3.2); this
is needed because pymssql still supports Python 2.7.  In the future,
this class may be removed and replaced with datetime.timezone.