Parse DATETIMEOFFSET into an aware datetime object.
The built-in DATETIMEOFFSET type is similar to DATETIME2, but
explicitly represents the offset between UTC and the local time zone,
so that timestamps can be manipulated sensibly even if they originate
from different time zones (or the same time zone at different times of
the year.)

The natural equivalent in Python is an "aware" datetime object, i.e. a
timestamp with an associated "tzinfo" object indicating the time zone.

Thus, when a DATETIMEOFFSET value is retrieved from the database
(which is only possible if we are using TDS 7.3 or later), convert it
into an aware datetime object.

Since pymssql still supports Python 2.7, which lacks the built-in
datetime.timezone class, the MSSQLTimezone class is used instead.

As with DATETIME2, parsing a DATETIMEOFFSET value, without loss of
precision, requires using the dbanydatecrack function rather than
dbdatecrack.

As a result of this change:

  >>> cursor.execute(
  ...     "SELECT CAST('2000-01-01 12:34:56.789123 -05:00' AS DATETIMEOFFSET)")
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 789123, tzinfo=MSSQLTimezone(datetime.timedelta(days=-1, seconds=68400))),)]

whereas previously:

  >>> cursor.execute(
  ...     "SELECT CAST('2000-01-01 12:34:56.789123 -05:00' AS DATETIMEOFFSET)")
  >>> cursor.fetchall()
  [(b'\x1e\x89\xd3_\x93\x00\x00\x00\xac\x8e\x00\x00\xd4\xfe\x07\xe0',)]