Preserve microseconds and timezone of datetime parameters.
If the TDS version is 7.3 or later, convert datetime parameters to
DATETIME2 or DATETIMEOFFSET, as appropriate.  Don't do this for TDS
7.2 and earlier; both for symmetry and for backward compatibility.

The historical behavior of pymssql is that when a datetime object is
used as a query parameter, it is serialized as something like:

  {ts '2000-01-01 12:34:56.789'}

which is more-or-less equivalent to:

  CAST('2000-01-01 12:34:56.789' AS DATETIME)

In particular, the "ts" format:
- forbids specifying more than 3 digits after the decimal point;
- forbids specifying a time zone offset;
- implicitly converts the value to DATETIME; i.e., the value is
  rounded to the nearest 1/300 second.

Moreover, note that the historical behavior is to truncate values to
the next lower millisecond, rather than rounding to the nearest.  So a
value such as datetime.datetime(2000, 1, 1, 12, 34, 56, 8999) is
converted to "{ts '2000-01-01 12:34:56.008'}", which is interpreted by
the server as 2000-01-01 12:34:56.0066667.

Here, we maintain the historical behavior as long as TDS 7.2 or
earlier is used.  If TDS 7.3 or newer is used, values are instead
serialized as something like:

  CAST('2000-01-01 12:34:56.789123' AS DATETIME2)

or:

  CAST('2000-01-01 12:34:56.789123-05:00' AS DATETIMEOFFSET)

(Note that SQL Server permits implicit conversions between all of
these types; for example, inserting either of these expressions into a
DATETIME column gives the same result as the "ts" expression above.)

As a result of this change, when using TDS 7.3:

  >>> cursor.execute("SELECT %s",
  ...                datetime.datetime(2000, 1, 1, 12, 34, 56, 8999))
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 8999),)]

  >>> cursor.execute("SELECT %s",
  ...     datetime.datetime(2000, 1, 1, 12, 34, 56, 8999,
  ...         datetime.timezone(datetime.timedelta(hours=1))))
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 8999, tzinfo=MSSQLTimezone(datetime.timedelta(days=seconds=3600))),)]

Previously, or when using TDS 7.2:

  >>> cursor.execute("SELECT %s",
  ...     datetime.datetime(2000, 1, 1, 12, 34, 56, 8999))
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 7000),)]

  >>> cursor.execute("SELECT %s",
  ...     datetime.datetime(2000, 1, 1, 12, 34, 56, 8999,
  ...         datetime.timezone(datetime.timedelta(hours=1))))
  >>> cursor.fetchall()
  [(datetime.datetime(2000, 1, 1, 12, 34, 56, 7000),)]

This has the desirable properties that:

- clients that are talking to old versions of SQL Server continue
  working as before;

- clients that explicitly select TDS 7.2 or earlier continue working
  as before;

- clients that are talking to new versions of SQL Server, using a new
  protocol version, obtain consistent and intuitive behavior.

It has some drawbacks:

- time zone information is (still) silently lost when using TDS 7.2 or
  earlier;

- clients that automatically select the TDS protocol version will get
  different rounding behavior depending on the server version;

- clients that explicitly select TDS 7.3 will get different (better)
  rounding behavior than they had with pymssql 2.1.4.