Detect newer MS SQL Server versions.
The internal function get_sql_server_version is used by some test
cases to check which actual version of SQL Server is in use.

Previously (due to a typo), for versions newer than SQL Server 2012,
this function would return 2000.  Instead, return 2014 in that case.

This function still ought to be updated to recognize the version codes
for newer server versions.