Test insert-and-select of aware datetime.datetime.
Test that it is possible to insert an aware datetime.datetime value as
a query parameter, and that the value (including the timezone offset,
and including microsecond precision) survives a round trip.

Furthermore, test that Python's interpretation of timezone offsets
matches the interpretation on the server: using Python's "astimezone"
to translate the value into a different timezone should yield a
DATETIMEOFFSET value that is considered "equal" by the server.