Add tds_version_tuple parameter to substitute_params.
This can be used to vary the formatting of query parameters based on
the active TDS protocol version.

Each of the following internal functions now accepts an optional
tds_version_tuple argument:

 _substitute_params (called by MSSQLConnection.format_sql_command)
 _quote_data (called by _substitute_params)
 _quote_or_flatten (called by _quote_data)
 _quote_simple_value (called by _quote_or_flatten and _quote_data)

(_quote_data contains an apparent bug, in that the charset is ignored
for simple values.  This is not addressed here.)